Rename batch_sys_manager -> job_runner_mgr
