+++ cylc/flow/scripts/jobs_poll.py
@@ -22,7 +22,7 @@ Read job status files to obtain the statuses of the jobs. If necessary, Invoke
 the relevant job runner commands to ask the job runners for more statuses.
 
 """
-from cylc.flow.batch_sys_manager import BatchSysManager
+from cylc.flow.job_runner_mgr import BatchSysManager
 from cylc.flow.option_parsers import CylcOptionParser as COP
 from cylc.flow.terminal import cli_function
